Fix TypeScript compatibility
